+++ cmd/users/main.go
@@ -9,58 +9,41 @@ import (
 	"time"
 
 	"github.com/gofiber/fiber/v2"
+	"github.com/italoservio/braz_ecommerce/cmd/users/start"
 	"github.com/italoservio/braz_ecommerce/packages/database"
 )
-
-type Healthcheck struct {
-	Server   string `json:"server"`
-	Database string `json:"database"`
-}
 
 func main() {
 	app := fiber.New()
 
-	dbUri := os.Getenv("DB_URI")
-	dbName := os.Getenv("DB_NAME")
+	env := start.NewEnv()
 
-	db, err := database.NewDatabase(dbUri, dbName)
+	db, err := database.NewDatabase(env.DB_URI, env.DB_NAME)
 	if err != nil {
 		log.Fatal(err)
 	}
 
-	app.Get("/health", healthcheck(db))
+	userController := start.InjectionsContainer(db)
 
-	go func() {
-		log.Fatal(app.Listen(":3000"))
-	}()
+	app.Get("/health", start.HealthCheckEndpoint(db))
+
+	api := app.Group("/api")
+	usersV1 := api.Group("/v1/users")
+	usersV1.Get("/:id", userController.GetUserById)
+
+	go func() { log.Fatal(app.Listen(env.PORT)) }()
 
 	stop := make(chan os.Signal, 1)
 	signal.Notify(stop, syscall.SIGINT, syscall.SIGTERM)
 	<-stop
 
+	gracefulShutdown(app, db)
+}
+
+func gracefulShutdown(app *fiber.App, db *database.Database) {
 	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
 	defer cancel()
 
 	app.ShutdownWithContext(ctx)
 	db.Client().Disconnect(ctx)
-}
-
-func healthcheck(db *database.Database) func(*fiber.Ctx) error {
-	return func(c *fiber.Ctx) error {
-		ctx, cancel := context.WithTimeout(context.Background(), time.Second*10)
-		defer cancel()
-
-		healthCheck := Healthcheck{
-			Server:   "healthy",
-			Database: "healthy",
-		}
-
-		err := db.Client().Ping(ctx, nil)
-
-		if err != nil {
-			healthCheck.Database = "offline"
-		}
-
-		return c.JSON(healthCheck)
-	}
 }